@@ -10,13 +10,13 @@
 	"devDependencies": {
 		"@sveltejs/adapter-auto": "^3.0.0",
 		"@sveltejs/adapter-static": "^3.0.1",
-		"@sveltejs/kit": "^2.0.0",
-		"@sveltejs/vite-plugin-svelte": "^3.0.0",
+		"@sveltejs/kit": "^2.5.27",
+		"@sveltejs/vite-plugin-svelte": "^4.0.0-next.6",
 		"autoprefixer": "^10.4.18",
 		"postcss": "^8.4.35",
-		"svelte": "^4.2.7",
+		"svelte": "^5.0.0",
 		"tailwindcss": "^3.4.1",
-		"vite": "^5.0.3"
+		"vite": "^5.4.4"
 	},
 	"type": "module"
 }
