@@ -1,8 +1,10 @@
 <script>
     import "../app.css";
     import Navbar from "../components/Navbar.svelte";
+
+    let { children } = $props();
 </script>
 
 <Navbar/>
 
-<slot />
+{@render children?.()}
